+++ tomcat/src/main/java/camp/nextstep/controller/IndexControllerFactory.java
@@ -2,16 +2,17 @@ package camp.nextstep.controller;
 
 import camp.nextstep.controller.strategy.NotFoundStrategy;
 import camp.nextstep.controller.strategy.RequestMethodStrategy;
+import org.apache.coyote.controller.ControllerFactory;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
 import java.util.List;
 
-public class ControllerIndexFactory implements ControllerFactory {
+public class IndexControllerFactory implements ControllerFactory {
 
     private final List<RequestMethodStrategy> requestMethods;
 
-    public ControllerIndexFactory(List<RequestMethodStrategy> requestMethods) {
+    public IndexControllerFactory(List<RequestMethodStrategy> requestMethods) {
         this.requestMethods = requestMethods;
     }
