+++ tomcat/src/main/java/camp/nextstep/controller/RegisterControllerFactory.java
@@ -2,19 +2,16 @@ package camp.nextstep.controller;
 
 import camp.nextstep.controller.strategy.NotFoundStrategy;
 import camp.nextstep.controller.strategy.RequestMethodStrategy;
-import camp.nextstep.controller.strategy.UnauthorizedStrategy;
-import org.apache.coyote.http11.exception.UnAuthorizedException;
+import org.apache.coyote.controller.ControllerFactory;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
 import java.util.List;
 
-public class ControllerLoginFactory implements ControllerFactory {
-
+public class RegisterControllerFactory implements ControllerFactory {
     private final List<RequestMethodStrategy> requestMethods;
-    private final UnauthorizedStrategy redirectStrategy = new UnauthorizedStrategy();
 
-    public ControllerLoginFactory(List<RequestMethodStrategy> requestMethods) {
+    public RegisterControllerFactory(List<RequestMethodStrategy> requestMethods) {
         this.requestMethods = requestMethods;
     }
 
@@ -25,11 +22,6 @@ public class ControllerLoginFactory implements ControllerFactory {
                 .findFirst()
                 .orElseGet(NotFoundStrategy::new);
 
-        try {
-            return requestMethodStrategy.serve(httpRequest);
-        } catch (UnAuthorizedException e) {
-            return redirectStrategy.serve(httpRequest);
-        }
-
+        return requestMethodStrategy.serve(httpRequest);
     }
 }
